requirements.txt
sentence-transformers
faiss-cpu
groq
python-dotenv